Improve Values() and Names() methods; test them.
Analysis of coverage data provided by Paul Johnson++ at cpancover.com
indicated that Data::Dumper's coverage was surprisingly low:

                                 stmt   bran   cond    sub
  blib/lib/Data/Dumper.pm        85.2   76.7   69.6   70.7

Herewith I am adding files to test the previously untested Names() and
Values() methods.

However, as might be expected, the fact that these two methods were previously
untested meant that problems with the methods' code were hidden and only
uncovered when I started to write tests on the basis of the coverage data.
For both of these methods, the only sensible type of argument is an array
reference.  Any other type of argument should be deemed an error.  Error
handling was added and tested.